Sounds added
There are now zombie sounds and reload sounds.

There is a bug with the reload sound playing while the game is paused.
Also a bug that allows the player to go through walls.
Player also randomly spins or turns inconsistently.

Will be working on these.

Game is on google drive.
https://drive.google.com/drive/folders/1_diNID0iqeDdOPsUlXqNezbFLa7dLldy?usp=sharing